Updated for MC 1.11.2
major changes were the item name, made it more consistent with vanilla (lower underscore separated).

Also tweaked the behaviour of the bench while holding to make it a bit more like the shield to hopefully stop some of the double hand bobbling, however it looks like some of that still exists with the shield as it sits.